+++ Angular2Emulator/TypeScript/Ng2Emulation/Templates/Rules/NgPropertyRule.ts
@@ -4,7 +4,7 @@ export default class NgPropertyRule implements ParserRule {
     processTemplate(template: string): string {
 
 		// # symbol to declare local variables
-		let regex = /<? #([a-zA-Z0-9-]+).+?>/g;
+		let regex = /<? #([a-zA-Z0-9-$]+).+?>/g;
 
         template = template.replace(regex, (text, match) => {
 	         return text.replace("#" + match, `ng-property="${match}"`);